% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_borgbjerg.R
\docType{data}
\name{Borgbjerg_m}
\alias{Borgbjerg_m}
\title{Aortic diameter data with multiple measurements per observer}
\format{A data frame with 1200 observations and 4 variables:
\describe{
  \item{subject}{aorta ID}
  \item{observer}{observer ID}
  \item{value}{aorta diameter, in mm}
  \item{measurement}{repetition ID, 1/2 = first/second measurement made by the specific observer on the specific aorta}
}}
\usage{
data(Borgbjerg_m)
}
\description{
Data frame containing measurements of aortic diameters obtained by the ITI method, see \insertCite{borgbjerg;textual}{loamr}.
For 50 aortas the diameter have been measured twice by 12 observers, giving a total of 1200 observations.
}
\examples{
data(Borgbjerg_m)

}
\references{
\insertRef{borgbjerg}{loamr}
}
\keyword{datasets}
