% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Erasmus.R
\docType{data}
\name{Erasmus}
\alias{Erasmus}
\title{Data from \insertRef{Erasmus2003}{loamr}}
\format{A data frame with 40 rows and 5 columns, describing tumour size on 40 tumours each measured by 5 readers.}
\source{
\href{http://?????.html}{Blog}
}
\usage{
data(Erasmus)
}
\description{
Data from \insertRef{Erasmus2003}{loamr}
}
\examples{
data(Erasmus)


}
\references{
\insertRef{Erasmus2003}{loamr}
}
\keyword{datasets}
