% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Jones.R
\docType{data}
\name{Jones}
\alias{Jones}
\title{Simulated data jones}
\format{A data frame with 9882 rows and 7 variables:}
\source{
\href{http://?????.html}{Blog}
}
\usage{
data(Jones)
}
\description{
Simulated data jones
}
\examples{
data(Jones)


}
\references{
\insertRef{Jones2011}{loamr}
}
\keyword{datasets}
