% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{print.loamobject}
\alias{print.loamobject}
\title{Print output from 'LOAM' function}
\usage{
\method{print}{loamobject}(x, digits = 3, ...)
}
\arguments{
\item{x}{an object of class 'loamobject'}

\item{digits}{number of digits shown in printed output}

\item{...}{passthrough}
}
\description{
Print output from 'LOAM' function
}
