% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMD.R
\name{simMD}
\alias{simMD}
\title{Simulates data from a two-way random effect model}
\usage{
simMD(
  mu = 0,
  sigma2A = 2,
  sigma2B = 1,
  sigma2E = 0.5,
  sigma2AB = NULL,
  interaction = F,
  n_subjects = 15,
  n_observers = 20,
  n_measurements = 1,
  n_sim = 1
)
}
\arguments{
\item{mu}{overall mean}

\item{sigma2A}{inter-subject variance}

\item{sigma2B}{inter-observer variance}

\item{sigma2E}{residual variance}

\item{sigma2AB}{subject-observer interaction variance, only used if 'interaction = T'}

\item{interaction}{logical, indicates if subject-observer interaction should be included in the two-way random effects model}

\item{n_subjects}{number of subjects}

\item{n_observers}{number of observers}

\item{n_measurements}{number of measurements per subject-observer pair}

\item{n_sim}{number of simulated datasets}
}
\value{
A tibble of simulated measurements. The tibble is in the format
required for the 'LOAM'-function when n_sim = 1, i.e. in long format with
columns 'subject', 'observer', 'measurement' (if h > 1), and 'value' (= the
simulated data). When n_sim > 1, the outputted tibble contains a column for
each of the simulated data sets, named 'value1', 'value2', etc.
}
\description{
Simulates data from a two-way random effect model given by the
formula A + B, where A denotes subject and B observer, as in
\insertCite{christensen;textual}{loamr}.
An interaction term can also be included, i.e. A + B + AB (as considered
in \insertCite{christensen2025;textual}{loamr}.
}
\examples{
simMD()

LOAM(simMD())

}
\references{
\insertRef{christensen}{loamr}
\insertRef{christensen2025}{loamr}
}
