% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_borgbjerg.R
\docType{data}
\name{Borgbjerg}
\alias{Borgbjerg}
\title{Single measurement data with the ITT meahod of}
\format{A data frame with 900 observations}
\usage{
data(Borgbjerg)
}
\description{
Single measurement data with the ITT meahod of
}
\examples{
data(Borgbjerg)

}
\references{
\insertRef{borgbjerg}{loamr}
}
\keyword{datasets}
