% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_borgbjerg.R
\docType{data}
\name{Borgbjerg_m}
\alias{Borgbjerg_m}
\title{Multiple measurement data with the ITI method of}
\format{A data frame with 1200 observations}
\usage{
data(Borgbjerg_m)
}
\description{
Multiple measurement data with the ITI method of
}
\examples{
data(Borgbjerg_m)

}
\references{
\insertRef{borgbjerg}{loamr}
}
\keyword{datasets}
