% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jones.R
\docType{data}
\name{Jones}
\alias{Jones}
\title{Data from Jones et al. (2011)}
\format{A data frame with 9882 rows and 7 variables:}
\source{
\href{http://?????.html}{Blog}
}
\usage{
data(Jones)
}
\description{
Data from Jones et al. (2011)
}
\examples{
data(Jones)


}
\references{
\insertRef{Jones2011}{loamr}
}
\keyword{datasets}
