% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOAM.R
\name{LOAM}
\alias{LOAM}
\title{Function LOAM 18-12-19}
\usage{
LOAM(data, CI = 0.95)
}
\arguments{
\item{data}{A dataframe of measurement data}

\item{CI}{Confidence interval on the LoAM,}
}
\value{
An object of class "ccobject".
}
\description{
This function calculaates the limit of agreement
with the mean. It calculates both symetrical and asymetrical
confidence intervals
}
\details{
The data argument requires data in long format with all
columns spelled lower case as here:

- subject :
- observer : an id of the individual for the specific measurement
- value : value of the measurement
}
\examples{
data(Jones)

Jones$subject = 1:nrow(Jones)
Jones <- tidyr::gather(Jones, observer, value, -subject)

LOAM(Jones)

}
