% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOAM.R
\name{LOAM}
\alias{LOAM}
\title{Calculates estimate of ...}
\usage{
LOAM(data, CI = 0.95)
}
\arguments{
\item{data}{A dataframe,}

\item{CI}{confidence interval on the LoAM,}
}
\value{
An object of class "ccobject".
}
\description{
This function...
}
\details{
The function returns ...
}
\examples{
data(Jones)

Jones$subject = 1:nrow(Jones)
Jones <- tidyr::gather(Jones, observer, value, -subject)

LOAM(Jones)

}
