% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOAM.R
\name{LOAM}
\alias{LOAM}
\title{Function LOAM 13-01-20}
\usage{
LOAM(data, CI = 0.95)
}
\arguments{
\item{data}{a data frame containing measurement data in long format (see 'Details')}

\item{CI}{coverage probability for the confidence interval on the LOAM.}
}
\value{
An object of class "loamobject".
}
\description{
This function calculates estimates and confidence intervals for the
95\% limits of agreement with the mean presented by \insertCite{christensen;textual}{loamr}.
}
\details{
The data argument requires data in long/narrow format with
the following columns:

- subject: a unique id for each subject

- observer: a unique id of the observer/reader

- value: value of the measurement

- measurement: an id indicating the measurement number if each observer
has performed multiple measurements on each subject. If only one measurement
per observer per subject, this column is not required.


The procedure requires balanced data, meaning that all observers must have measured
all subjects the same number of times.

The function outputs estimates and approximate CIs for the LOAM, the
intra-class correlation, \eqn{\sigma_A}, \eqn{\sigma_B}, and \eqn{\sigma_E}, where
\eqn{\sigma_A^2}, \eqn{\sigma_B^2}, and \eqn{\sigma_E^2} are the variance components of the underlying
two-way random model.  See \insertCite{christensen;textual}{loamr} for details.
}
\examples{
data(Borgbjerg)
L <- LOAM(Borgbjerg)
L

# Point estimates of LOAM, sigmaA, sigmaB, sigmaE, and ICC(A, 1):
L$estimates

# CIs for the LOAM, sigmaA, sigmaB, sigmaE, and ICC(A, 1):
L$intervals

}
\references{
\insertRef{christensen}{loamr}
}
