% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOAM.R
\name{LOAM}
\alias{LOAM}
\title{Function LOAM 13-01-20}
\usage{
LOAM(data, CI = 0.95)
}
\arguments{
\item{data}{A dataframe of measurement data}

\item{CI}{confidence interval on the LoAM,}
}
\value{
An object of class "loamobject".
}
\description{
This function calculaates the limit of agreement
with the mean. It calculates both symetrical and asymetrical
confidence intervals.
}
\details{
The data argument requires data in long/narrow format with all
columns spelled lower case, the required columns are the following:

- subject : a unique id for each subject

- observer : a unique id of the individual

- value : value of the measurement

- measurement : a unique id for each measurement, note that this is optional

The estimation requires balanced data, so if you want two measurements you need
that for all subject/observer combinations. This is also true for subjects or observers.
This means that all observers must have measured all subjects.
}
\examples{
data(Erasmus)

Erasmus$subject = 1:nrow(Erasmus)
Erasmus <- tidyr::gather(Erasmus, observer, value, -subject)

LOAM(Erasmus)

}
