% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOAM_ci_width.R
\name{LOAM_ci_width}
\alias{LOAM_ci_width}
\title{Width of CI for LOAM (for sample size calculation)}
\usage{
LOAM_ci_width(a, b, c, sigma2B_0, sigma2E_0, CI = 0.95)
}
\arguments{
\item{a}{number of subjects}

\item{b}{number of observers}

\item{c}{number of measurements per observer per subject}

\item{sigma2B_0}{observer variance}

\item{sigma2E_0}{residual variance}

\item{CI}{coverage probability for the confidence interval on the LOAM.}
}
\value{
Width of the CI for the 95\% LOAM for the selected parameter values.
}
\description{
Width of the confidence intervals for the
95\% limits of agreement with the mean for specified parameter values (\insertCite{christensen;textual}{loamr}).
}
\details{
Given initial estimates of the observer and residual variance
components (\code{sigma2B_0} and \code{sigma2E_0}, respectively) we can determine the width
of the CI for the 95\% LOAM for specified numbers of subjects, observers, and
measurements per observer per subject (\code{a}, \code{b}, and \code{c}, respectively)
(\insertCite{christensen;textual}{loamr}).

One of \code{a}, \code{b}, and \code{c} is allowed to be a vector of length
greater than one. Usually, \code{a} and \code{c} are fixed, while we want to
determine how many observers, \code{b}, we need to obtain a certain width of
the CI for the LOAM.
}
\examples{
LOAM_ci_width(a = 20, b = 4:10, c = 2, sigma2B_0 = 0.5, sigma2E_0 = 1)


}
\references{
\insertRef{christensen}{loamr}
}
