% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot.loamobject}
\alias{plot.loamobject}
\title{Adding to the generic plot function}
\usage{
\method{plot}{loamobject}(x, CItype = "asym", ...)
}
\arguments{
\item{x}{an object of class 'loamobject'}

\item{CItype}{confidence interval type: symmetric ("sym") or asymmetric ("asym")}

\item{...}{passthrough}
}
\value{
An agreement plot with an estimate and CI for the 95\% limits of agreement with the mean
}
\description{
Adding to the generic plot function
}
