% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot.loamobject}
\alias{plot.loamobject}
\title{Adding to the generic plot function}
\usage{
\method{plot}{loamobject}(x, CItype = "asym", ...)
}
\arguments{
\item{x}{The loamobject}

\item{CItype}{confidence interval type: "sym" or "asym"}

\item{...}{passthrough}
}
\value{
a plot
}
\description{
Adding to the generic plot function
}
