% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot.loamobject}
\alias{plot.loamobject}
\title{Adding to the generic plot function}
\usage{
\method{plot}{loamobject}(x, ...)
}
\arguments{
\item{x}{an object of class 'loamobject'}

\item{...}{passthrough}
}
\value{
An agreement plot with an estimate and CI for the 95\% reproducibility limits of agreement with the mean
}
\description{
Adding to the generic plot function
}
