% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{print.loamobject}
\alias{print.loamobject}
\title{Adding to the generic print function}
\usage{
\method{print}{loamobject}(x, ...)
}
\arguments{
\item{x}{The loamobject}

\item{...}{passthrough}
}
\value{
a plot
}
\description{
Adding to the generic print function
}
