% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMD.R
\name{simMD}
\alias{simMD}
\title{simMD function 18-12-19}
\usage{
simMD(
  subjects = 15,
  observers = 20,
  measurements = 1,
  mu = 0,
  sigma2A = 0.1,
  sigma2B = 0.1,
  sigma2E = 0.5
)
}
\arguments{
\item{subjects}{number of subjects}

\item{observers}{number of observers}

\item{measurements}{number of measurements}

\item{mu}{overall mean}

\item{sigma2A}{inter-subject variance}

\item{sigma2B}{inter-observer variance}

\item{sigma2E}{residual variance}
}
\value{
A dataframe of simulated measurements. The data frame is in the format
required for the 'LOAM'-function.
}
\description{
Simulates data from the random effect models described in
\insertCite{christensen;textual}{loamr}
}
\details{
The function returns a dataframe
}
\examples{
simMD()

LOAM(simMD())

}
\references{
\insertRef{christensen}{loamr}
}
