% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMD.R
\name{simMD}
\alias{simMD}
\title{simMD function, updated 08-07-2025}
\usage{
simMD(
  mu = 0,
  sigma2A = 2,
  sigma2B = 1,
  sigma2E = 0.5,
  sigma2AB = NULL,
  interaction = F,
  n_subjects = 15,
  n_observers = 20,
  n_measurements = 1
)
}
\arguments{
\item{mu}{overall mean}

\item{sigma2A}{inter-subject variance}

\item{sigma2B}{inter-observer variance}

\item{sigma2E}{residual variance}

\item{sigma2AB}{subject-observer interaction variance, only used if 'interaction = T'}

\item{interaction}{logical, indicates if subject-observer interaction should be included in the two-way random effects model}

\item{n_subjects}{number of subjects}

\item{n_observers}{number of observers}

\item{n_measurements}{number of measurements per subject-observer pair}
}
\value{
A tibble of simulated measurements. The tibble is in the format
required for the 'LOAM'-function.
}
\description{
Simulates data from a two-way random effect model given by the formula A + B, where A denotes subject and B observer, as in \insertCite{christensen;textual}{loamr}.
An interaction term can also be included, i.e. A + B + AB.
}
\examples{
simMD()

LOAM(simMD())

}
\references{
\insertRef{christensen}{loamr}
}
