% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMD.R
\name{simMD}
\alias{simMD}
\title{Simulates measurement data}
\usage{
simMD(subjects = 15, observers = 20, measurements = 1, mu = 0,
  sigma2A = 0.1, sigma2B = 0.1, sigma2E = 0.5)
}
\arguments{
\item{subjects}{A dataframe,}

\item{observers}{A dataframe,}

\item{measurements}{A dataframe,}

\item{mu}{A dataframe,}

\item{sigma2A}{A dataframe,}

\item{sigma2B}{A dataframe,}

\item{sigma2E}{A dataframe,}
}
\value{
A dataframe of simulated measurements
}
\description{
This function simulates measurement data
}
\details{
The function returns a dataframe
}
\examples{
simMD()

}
