% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMD.R
\name{simMD}
\alias{simMD}
\title{simMD function 18-12-19}
\usage{
simMD(
  subjects = 15,
  observers = 20,
  measurements = 1,
  mu = 0,
  sigma2A = 0.1,
  sigma2B = 0.1,
  sigma2E = 0.5
)
}
\arguments{
\item{subjects}{Number of subjects to simulate}

\item{observers}{Number of observers to simulate}

\item{measurements}{Number of measurements to simulate}

\item{mu}{Mean of simulation}

\item{sigma2A}{parameter}

\item{sigma2B}{parameter}

\item{sigma2E}{parameter}
}
\value{
A dataframe of simulated measurements
}
\description{
Simulates measurement data to test that can be
used to test coverage of the LOAM function.
}
\details{
The function returns a dataframe
}
\examples{
simMD()

}
